Fixed lower caps bug
Fixed correct caps filenames not being commited
