+++ client/src/components/Register/RegisterTransportador.vue
@@ -67,13 +67,6 @@
 
             <button type="submit" class="btn btn-primary" id="registerButton" style="width: 65%; margin-left: 17.5%;">Registar como Transportador&nbsp;&nbsp;<font-awesome-icon :icon="['fa', 'leaf']" size="lg"/></button>
 
-            <div class="or-seperator"><i>ou</i></div>
-            <p class="text-center">Regista-te através de uma rede social</p>
-            <div class="text-center social-btn">
-                <a href="#" class="btn btn-danger"><font-awesome-icon :icon="['fab', 'google']" size="lg"/>&nbsp; Google</a>
-                <a href="#" class="btn btn-secondary"><font-awesome-icon :icon="['fab', 'facebook-square']" size="lg"/>&nbsp; Facebook</a>
-            </div>
-
         </form>
         <p class="text-center text-muted small">Já tens conta?  <router-link to="/login" class="float-right">Inicia sessão aqui!</router-link></p>
     </div>
